requirements.txt
torch
torchvision
tqdm
numpy
pandas
scikit-learn
Pillow
streamlit
kaggle
matplotlib
opencv-python